+++ MainLib/IDictionary.h
@@ -13,7 +13,7 @@ namespace Hlib
     {
         virtual ET_ReturnCode eSetDbPath (const CEString& sDbPath) = 0;
         virtual CEString sGetDbPath() = 0;
-        virtual ET_ReturnCode eGetLexemeById(long Id) = 0;
+        virtual ET_ReturnCode eGetLexemeById(int Id) = 0;
         virtual ET_ReturnCode eGetLexemesByHash (const CEString& sMd5) = 0;
         virtual ET_ReturnCode eGetLexemesByGraphicStem (const CEString&) = 0;
         virtual ET_ReturnCode eGetLexemesByInitialForm (const CEString&) = 0;
